reorganized link module
